requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-dotenv==1.0.1
google-generativeai==0.7.2
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-multipart==0.0.9
pydantic==2.7.4
httpx==0.27.0
